+++ examples/data-sources/iosxr_evpn_evi/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxr_evpn_evi" "example" {
+  vpn_id = 1234
+}