+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add `rewrite_ingress_tag_two` attribute to `iosxr_interface` resource
 - Add `encapsulation_dot1q_vlan_id` attribute to `iosxr_interface` resource
 - Add `load_interval` attribute to `iosxr_interface` resource
+- Add `iosxr_evpn_evi` resource and data source
 
 ## 0.1.4